requirements.txt
aiogram==3.13.1
openai==1.54.4
python-dotenv==1.0.1
pydub==0.25.1
aiofiles==24.1.0
aiohttp==3.10.11
pytest==8.3.3
pytest-asyncio==0.24.0